requirements.txt
# Core Data Science
pandas>=2.1.4
numpy>=1.24.3
scikit-learn>=1.3.2
scipy>=1.11.4

# Machine Learning
xgboost>=2.0.0
tensorflow>=2.13.0
shap>=0.45.0
lime>=0.2.0.1

# Data Collection APIs
yfinance>=0.2.28
alpha-vantage>=2.3.1
fredapi>=0.5.0
requests>=2.31.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.12.2
plotly>=5.17.0

# Configuration
pyyaml>=6.0

# Development Tools
python-dotenv>=1.0.0
jupyter>=1.0.0
notebook>=7.0.6

# Testing
pytest>=7.4.3

# Web Application (for later weeks)
streamlit>=1.28.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0